requirements.txt
openai>=1.0.0
pinecone-client>=2.2.4
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-pinecone>=0.0.1
python-dotenv>=1.0.0
tenacity>=8.2.0 